Memoize the chord diagram renderer and inline error message in rendered markdown. (#33)
Memoize chord diagram and inline error messages.

This PR memoizes the chord diagram renderer to improve performance. If the chord diagram renderer is called with the same arguments as a previous call, the originally produced SVG will be cached and returned.

We also catch error messages from plugins and render the error message they throw.